requirements.txt
# Core ML / Embeddings
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.1

# Vector Search
faiss-cpu==1.8.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# API / Backend
flask==3.0.3

# Utilities
tqdm==4.66.4
requests==2.32.3

# Optional (if you used them)
python-dotenv==1.0.1